+++ src/main/java/com/example/tc/tech_challange/domain/pessoas/PessoasRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.web.bind.annotation.PathVariable;
 
-public interface PessoasRepository extends JpaRepository<Pessoas, Long> {
+public interface PessoasRepository extends JpaRepository<Pessoas, Integer> {
     Pessoas getReferenceById(PathVariable id);
 
     Page<Pessoas> findAll(Pageable paginacao);
